requirements.txt
# ==========================================
# CORE MACHINE LEARNING (PyTorch)
# ==========================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ==========================================
# TRANSFORMERS & HUGGING FACE
# ==========================================
# Pinned below v5.0.0 to prevent dependency conflicts on macOS
transformers<5.0.0
datasets>=2.14.0
sentencepiece>=0.1.99

# ==========================================
# COMPUTER VISION & IMAGE PROCESSING
# ==========================================
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.3

# ==========================================
# NLP & TEXT PROCESSING
# ==========================================
spacy>=3.7.0

# ==========================================
# DATA SCIENCE & METRICS
# ==========================================
numpy>=1.24.0
scikit-learn>=1.3.0